+++ engine/runtime/src/main/java/org/enso/interpreter/node/callable/function/BlockNode.java
@@ -9,6 +9,8 @@ import com.oracle.truffle.api.nodes.NodeInfo;
 import com.oracle.truffle.api.source.SourceSection;
 import java.util.Set;
 import org.enso.interpreter.node.ExpressionNode;
+import org.enso.interpreter.runtime.EnsoContext;
+import org.enso.interpreter.runtime.error.DataflowError;
 
 /**
  * This node defines the body of a function for execution, as well as the protocol for executing the
@@ -55,8 +57,15 @@ public class BlockNode extends ExpressionNode {
   @Override
   @ExplodeLoop
   public Object executeGeneric(VirtualFrame frame) {
+    var ctx = EnsoContext.get(this);
+    var nothing = ctx.getBuiltins().nothing();
     for (ExpressionNode statement : statements) {
-      statement.executeGeneric(frame);
+      var result = statement.executeGeneric(frame);
+      if (result != nothing) {
+        if (result instanceof DataflowError err) {
+          return err;
+        }
+      }
     }
     return returnExpr.executeGeneric(frame);
   }